Rework the parser to separate the frontend and backend better
Also added in the addition instructions
