+++ Makefile
@@ -4,7 +4,8 @@ EXEC = pun-vm
 OBJECTS = main.o parser/parser.o \
 					instructions/instructions.o \
 					instructions/store_memory/store_memory.o \
-					instructions/halting/halting.o
+					instructions/halting/halting.o \
+					instructions/addition/addition.o
 DEPENDS = ${OBJECTS:.o=.d}
 
 ${EXEC}: ${OBJECTS}
