+++ examples/test1.pun
@@ -2,7 +2,7 @@ slime-massacre 0, 0, 100
 laugh-wildly 1, 0, 0
 a-dry-dog 10, 0, 1
 peppy 10
-hufflepuff
+hufflepuff-aloud-light-tell
 
 qFirst line:
 qRegisters:
