+++ instructions/halting/halting.h
@@ -7,7 +7,6 @@ class Halting : public Instructions {
   public:
     Halting();
     void performInstruction() override;
-    uint32_t transformToInteger() override {return 0x1100;};
     virtual ~Halting() {};
 };
