+++ instructions/store_memory/store_memory.h
@@ -8,7 +8,6 @@ class StoreMemory : public Instructions {
     StoreMemory(int reg1, int reg2, int reg3);
     StoreMemory(const string &s);
     void performInstruction() override;
-    uint32_t transformToInteger() override {return 0x1100;};
     virtual ~StoreMemory() {};
     string toString() const override;
 };
