+++ main.cc
@@ -19,14 +19,11 @@ int main(int argc, char *argv[]) {
 		throw "Please enter a valid command!";
 	}
 
-	// Step 1: Transform .pun to binary, and load the binary into memory
+	// Step 1: Transform .pun to a bunch of structs	
 	// Step 2: Have a program counter to point to instruction we take out
 		// Then, we decode it (into a specific instruction)
 		// Then we execute that instruction
 	Parser::writePUNInstructionsToMemory(argv[1], memory);
-
-	// for testing
-	parseInstruction("slime-massacre-");
 
 	return 0;
 }
